+++ mobile/app/terminal-settings.tsx
@@ -8,7 +8,7 @@ import Animated, {
   useSharedValue
 } from 'react-native-reanimated'
 import { useRouter } from 'expo-router'
-import { ChevronLeft, ChevronRight, Smartphone } from 'lucide-react-native'
+import { ChevronLeft, ChevronRight, Smartphone, Type } from 'lucide-react-native'
 import { colors, radii, spacing, typography } from '../src/theme/mobile-theme'
 import { loadHosts } from '../src/transport/host-store'
 import type { HostProfile } from '../src/transport/types'
@@ -17,8 +17,33 @@ import type { RpcClient } from '../src/transport/rpc-client'
 import { PickerModal, type PickerOption } from '../src/components/PickerModal'
 import { TerminalShortcutSettings } from '../src/components/TerminalShortcutSettings'
 import { setTerminalAutoRestoreFitMsForHost } from '../src/terminal/terminal-auto-restore-fit-state'
+import {
+  loadTerminalTextScale,
+  saveTerminalTextScale
+} from '../src/storage/preferences'
 
 type RestoreValue = 'indefinite' | '60s' | '5m' | '30m'
+
+type TextSizeValue = 'smallest' | 'smaller' | 'default' | 'large' | 'larger' | 'largest'
+
+// scale = baseline zoom the terminal WebView applies on top of fit-to-width.
+// Keep in sync with TERMINAL_TEXT_SCALES; pinch-to-zoom snaps to these values.
+const TEXT_SIZE_OPTIONS: (PickerOption<TextSizeValue> & { scale: number })[] = [
+  { value: 'smallest', label: 'Smallest (50%)', scale: 0.5 },
+  { value: 'smaller', label: 'Smaller (75%)', scale: 0.75 },
+  { value: 'default', label: 'Default (100%)', scale: 1 },
+  { value: 'large', label: 'Large (125%)', scale: 1.25 },
+  { value: 'larger', label: 'Larger (150%)', scale: 1.5 },
+  { value: 'largest', label: 'Largest (200%)', scale: 2 }
+]
+
+function textSizeValueFromScale(scale: number): TextSizeValue {
+  return TEXT_SIZE_OPTIONS.find((o) => o.scale === scale)?.value ?? 'default'
+}
+
+function textSizeSummary(scale: number): string {
+  return (TEXT_SIZE_OPTIONS.find((o) => o.scale === scale) ?? TEXT_SIZE_OPTIONS[0]!).label
+}
 
 const AUTO_RESTORE_FIT_OPTIONS: (PickerOption<RestoreValue> & { ms: number | null })[] = [
   { value: 'indefinite', label: 'Keep at phone size (default)', ms: null },
@@ -113,6 +138,20 @@ export default function TerminalSettingsScreen() {
   // drawer appear cut-off.
   const [hostMs, setHostMs] = useState<Record<string, number | null | undefined>>({})
   const [pickerHostId, setPickerHostId] = useState<string | null>(null)
+
+  const [textScale, setTextScale] = useState(1)
+  const [textSizePickerOpen, setTextSizePickerOpen] = useState(false)
+  useEffect(() => {
+    void loadTerminalTextScale().then(setTextScale)
+  }, [])
+  const selectTextSize = useCallback((value: TextSizeValue) => {
+    const opt = TEXT_SIZE_OPTIONS.find((o) => o.value === value)
+    if (!opt) {
+      return
+    }
+    setTextScale(opt.scale)
+    void saveTerminalTextScale(opt.scale)
+  }, [])
 
   useEffect(() => {
     let cancelled = false
@@ -243,6 +282,27 @@ export default function TerminalSettingsScreen() {
           </View>
         )}
 
+        <Text style={[styles.groupHeading, styles.inputGroupGap]}>TEXT SIZE</Text>
+        <Text style={styles.groupDescription}>
+          Scale the terminal text. Smaller sizes fit more columns with side margins; larger sizes
+          show fewer columns — drag sideways to pan. You can also pinch to zoom in the terminal
+          itself, which updates this setting. Per-device display only; doesn&apos;t change the
+          desktop terminal.
+        </Text>
+        <View style={[styles.section, styles.sectionTopGap]}>
+          <Pressable
+            style={({ pressed }) => [styles.row, pressed && styles.rowPressed]}
+            onPress={() => setTextSizePickerOpen(true)}
+          >
+            <Type size={16} color={colors.textSecondary} />
+            <View style={styles.rowContent}>
+              <Text style={styles.rowLabel}>Text size</Text>
+              <Text style={styles.rowSublabel}>{textSizeSummary(textScale)}</Text>
+            </View>
+            <ChevronRight size={16} color={colors.textMuted} />
+          </Pressable>
+        </View>
+
         <TerminalShortcutSettings
           scrollRef={scrollRef}
           scrollOffsetY={scrollOffsetY}
@@ -262,6 +322,15 @@ export default function TerminalSettingsScreen() {
           }
         }}
         onClose={() => setPickerHostId(null)}
+      />
+
+      <PickerModal<TextSizeValue>
+        visible={textSizePickerOpen}
+        title="Terminal text size"
+        options={TEXT_SIZE_OPTIONS}
+        selected={textSizeValueFromScale(textScale)}
+        onSelect={selectTextSize}
+        onClose={() => setTextSizePickerOpen(false)}
       />
     </GestureHandlerRootView>
   )
@@ -317,6 +386,9 @@ const styles = StyleSheet.create({
   },
   sectionTopGap: {
     marginTop: spacing.sm
+  },
+  inputGroupGap: {
+    marginTop: spacing.xl
   },
   emptyText: {
     fontSize: typography.bodySize,
