+++ mobile/src/storage/preferences.ts
@@ -25,6 +25,37 @@ export async function savePushNotificationsEnabled(enabled: boolean): Promise<vo
   await AsyncStorage.setItem(NOTIF_KEY, String(enabled))
 }
 
+const TEXT_SCALE_KEY = 'orca:terminalTextScale'
+
+// Why: the mobile terminal fits the desktop's full column count to the phone
+// width with a CSS scale, so xterm's raw fontSize is cancelled out and can't
+// drive apparent size. Instead we persist a baseline zoom multiplier ("text
+// size") that the WebView applies on top of the fit. Discrete presets keep the
+// settings picker simple and bound the value to ones the zoom logic handles;
+// pinch-to-zoom in the terminal snaps to these same presets. Sub-1 steps shrink
+// below fit-to-width (more columns visible with side margins).
+export const TERMINAL_TEXT_SCALES = [0.5, 0.75, 1, 1.25, 1.5, 2] as const
+const DEFAULT_TEXT_SCALE = 1
+
+export async function loadTerminalTextScale(): Promise<number> {
+  try {
+    const raw = await AsyncStorage.getItem(TEXT_SCALE_KEY)
+    if (raw === null) {
+      return DEFAULT_TEXT_SCALE
+    }
+    const parsed = Number(raw)
+    return (TERMINAL_TEXT_SCALES as readonly number[]).includes(parsed)
+      ? parsed
+      : DEFAULT_TEXT_SCALE
+  } catch {
+    return DEFAULT_TEXT_SCALE
+  }
+}
+
+export async function saveTerminalTextScale(scale: number): Promise<void> {
+  await AsyncStorage.setItem(TEXT_SCALE_KEY, String(scale))
+}
+
 export type HostPreferences = {
   sortMode: string
   filterMode: string
